pyproject.toml
[project]
name = "safety-checker"
version = "0.1.0"
description = "Integrate stable diffusion safety checker / NSFW image detection into web service like FastAPI."
authors = [{ name = "iyume", email = "iyumelive@gmail.com" }]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "Apache-2.0" }
dependencies = ["Pillow>=8.4.0"]

[project.optional-dependencies]
sdmock = ["transformers>=4.0", "diffusers[torch]>=0.20"]


[tool.pyright]
typeCheckingMode = "basic"
reportUnknownParameterType = true # cannot check return...

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
